requirements.txt
beautifulsoup4
boto3
colorlog
configparser
keyring
PyYAML
requests
pyotp
botocore